requirements.txt
Flask==2.3.2
python-docx==0.8.11
pdfplumber==0.10.2
pytesseract==0.3.10
Pillow==9.5.0
nltk==3.8.1